Moves utils tests
